Use the common PRNG in 2D particles code.
Replaces the custom PRNG used by 2D particles code with a wrapper for the PRNG located under core/math.

+++ scene/2d/particles_2d.h
@@ -127,7 +127,7 @@ private:
 		Vector2 velocity;
 		float rot;
 		float frame;
-		uint32_t seed;
+		uint64_t seed;
 		Particle() { active=false; seed=123465789; rot=0; frame=0;}
 	};
